consolidate common extensions API stuff to /browser
- remove the specific foreground environment implementations (and hence the webpack alias)
- move browser-specific API impls (and migration, etc.) to individual background.entry.js files (instead of other browsers importing from Chrome)
- remove the license comment (since it'll get buried in the middle of the bundle) and include the LICENSE file separately
