+++ browser/index.js
@@ -1,0 +1,8 @@
+/* @flow */
+
+export {
+	sendMessage,
+	sendSynchronous,
+	addListener,
+	addInterceptor,
+} from './foreground';
