+++ browser/utils/__tests__/messaging.js
