+++ chrome/manifest.json
@@ -11,7 +11,7 @@
 		]
 	},
 	"homepage_url": "http://redditenhancementsuite.com",
-	"options_page": "{{./options.html}}",
+	"options_page": "{{../browser/options.html}}",
 	"content_scripts": [
 		{
 			"matches": [
