+++ firefox/nativeRequire.js
