+++ lib/environment/ajax.js
@@ -1,9 +1,9 @@
 /* @flow */
 
 import _ from 'lodash';
+import { sendMessage } from '../../browser';
 import { loggedInUserHash, string } from '../utils';
 import { XhrCache } from './';
-import { sendMessage } from 'browserEnvironment';
 
 type ResponseType = 'text' | 'json' | 'raw';
