+++ lib/environment/alert.js
@@ -1,6 +1,6 @@
 /* @flow */
 
+import { addListener } from '../../browser';
 import { Alert } from '../utils';
-import { addListener } from 'browserEnvironment';
 
 addListener('alert', text => { Alert.open(text); });
