+++ lib/environment/auth.js
@@ -1,7 +1,7 @@
 /* @flow */
 
+import { sendMessage } from '../../browser';
 import { getUrlParams, insertParams, randomHash } from '../utils';
-import { sendMessage } from 'browserEnvironment';
 
 type AuthOptions = {
 	domain: string,
