+++ lib/environment/history.js
@@ -1,7 +1,7 @@
 /* @flow */
 
+import { sendMessage } from '../../browser';
 import { isPrivateBrowsing } from './';
-import { sendMessage } from 'browserEnvironment';
 
 export async function addURLToHistory(url: string): Promise<void> {
 	if (await isPrivateBrowsing()) return;
