+++ lib/environment/multicast.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { addListener, sendMessage } from 'browserEnvironment';
+import { addListener, sendMessage } from '../../browser';
 
 const callbacks = new Map();
