+++ lib/environment/privateBrowsing.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import _ from 'lodash';
-import { sendMessage } from 'browserEnvironment';
+import { sendMessage } from '../../browser';
 
 /**
  * @returns {Promise<boolean, *>}
