+++ lib/environment/session.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from 'browserEnvironment';
+import { sendMessage } from '../../browser';
 
 export function get(key: string): Promise<any | void> {
 	return sendMessage('session', ['get', key]);
