+++ lib/environment/storage.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import _ from 'lodash';
-import { sendMessage } from 'browserEnvironment';
+import { sendMessage } from '../../browser';
 
 export function get(key: string): Promise<any | null> {
 	return sendMessage('storage', ['get', key]);
