+++ lib/environment/tabs.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from 'browserEnvironment';
+import { sendMessage } from '../../browser';
 
 export function openNewTab(url: string, focus?: boolean = true) {
 	return openNewTabs(focus, url);
