+++ lib/environment/xhrCache.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { sendMessage } from 'browserEnvironment';
+import { sendMessage } from '../../browser';
 
 export function set(key: string, value: mixed) {
 	return sendMessage('XHRCache', ['set', key, value]);
